api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
libvirt-python==9.0.0
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
psutil==5.9.6
aiohttp==3.9.1
cryptography==41.0.7
pyOpenSSL==23.3.0
